refactor: ui -> controller로 네이밍 변경
