+++ src/main/java/user/controller/ListUserController.java
@@ -1,11 +1,10 @@
-package user.ui;
+package user.controller;
 
 import user.view.UserView;
-import webserver.RequestHandler;
+import webserver.controller.AbstractController;
 import webserver.domain.Body;
 import webserver.domain.HttpRequest;
 import webserver.domain.HttpResponse;
-import webserver.ui.AbstractController;
 
 public class ListUserController extends AbstractController {
     @Override
