+++ src/main/java/user/controller/LoginController.java
@@ -1,10 +1,10 @@
-package user.ui;
+package user.controller;
 
 import user.model.User;
 import user.service.UserService;
+import webserver.controller.AbstractController;
 import webserver.domain.HttpRequest;
 import webserver.domain.HttpResponse;
-import webserver.ui.AbstractController;
 
 public class LoginController extends AbstractController {
     @Override
