+++ src/main/java/webserver/controller/Controller.java
@@ -1,4 +1,4 @@
-package webserver.ui;
+package webserver.controller;
 
 import webserver.domain.HttpRequest;
 import webserver.domain.HttpResponse;
